开发 忽略mac语言包

+++ electron-builder.config.js
@@ -181,19 +181,17 @@ let build = {
         differentialPackage: false,
     },
     afterPack: async (c) => {
-        const localsPath = path.join(
-            c.appOutDir,
-            process.platform === "darwin" ? "e-search.app/Contents/Locales" : "locales"
-        );
-        try {
-            const files = fs.readdirSync(localsPath).filter((i) => i != "zh-CN.pak" && i != "en-US.pak");
-            for (let i of files) {
-                fs.rmSync(path.join(localsPath, i));
+        const localsPath = path.join(c.appOutDir, "locales");
+        if (process.platform != "darwin")
+            try {
+                const files = fs.readdirSync(localsPath).filter((i) => i != "zh-CN.pak" && i != "en-US.pak");
+                for (let i of files) {
+                    fs.rmSync(path.join(localsPath, i));
+                }
+                console.log("移除原生语言包");
+            } catch (error) {
+                console.log(error);
             }
-            console.log("移除原生语言包");
-        } catch (error) {
-            console.log(error);
-        }
 
         const appPath = path.join(
             c.appOutDir,
